+++ lib/scout_apm/capacity.rb
@@ -4,7 +4,7 @@ class ScoutApm::Capacity
 
 	def initialize
 		@processing_start_time = Time.now
-		@lock ||= Mutex.new
+		@lock ||= Mutex.new # the transaction_entry_time could be modified while processing a request or when #process is called.
 		@accumulated_time = 0.0
 	end
 	
@@ -20,15 +20,17 @@ class ScoutApm::Capacity
 		@lock.synchronize do
 			@accumulated_time += (Time.now - transaction_entry_time).to_f
 			@transaction_entry_time = nil
+			ScoutApm::Agent.instance.logger.debug "Accumulated time spent process requests: #{accumulated_time}"
 		end
 	end
 
 	# Ran when sending metrics to server. Reports capacity usage metrics.
 	def process
 		process_time = Time.now
+		ScoutApm::Agent.instance.logger.debug "Processing capacity usage for [#{@processing_start_time}] to [#{process_time}]. Time Spent: #{@accumulated_time}."
 		@lock.synchronize do
 			time_spent = @accumulated_time
-			@accumulated_time = 0
+			@accumulated_time = 0.0
 			# If a transaction is still running, capture its running time up to now and 
 			# reset the +transaction_entry_time+ to now. 
 			if @transaction_entry_time