+++ lib/scout_apm/tracer.rb
@@ -23,9 +23,13 @@ module ScoutApm::Tracer
       ScoutApm::Context.current.add_user(:ip => options[:ip]) if options[:ip]    
       Thread::current[:scout_apm_trace_time] = Time.now.utc
       ScoutApm::Agent.instance.capacity.start_transaction!
+      e = nil
       instrument(metric_name, options) do
         Thread::current[:scout_apm_scope_name] = metric_name
-        yield
+        begin
+          yield
+        rescue Exception => e
+        end
         Thread::current[:scout_apm_scope_name] = nil
       end
       Thread::current[:scout_apm_trace_time] = nil
@@ -33,6 +37,7 @@ module ScoutApm::Tracer
       # The context is cleared after instrumentation (rather than before) as tracing controller-actions doesn't occur until the controller-action is called.
       # It does not trace before filters, which is a likely spot to add context. This means that any context applied during before_filters would be cleared.
       ScoutApm::Context.clear!
+      raise e if e
     end
     
     # Options: